S.Res. 217 Introduced in Senate
A resolution expressing the sense of the Senate that Secretary of Health and Human Services Robert Fitzgerald Kennedy Jr. does not have the confidence of the Senate or of the American people to faithfully carry out the duties of his office.

Version:  Introduced in Senate (is)
Date:     2025-05-12
Congress: 119

Source: https://www.govinfo.gov/content/pkg/BILLS-119sres217is/xml/BILLS-119sres217is.xml

Sponsored-By: A000382
Cosponsor-Count: 3