H.Res. 426 Reported in House
Providing for consideration of the joint resolution (S.J. Res. 13) providing for congressional disapproval under chapter 8 of title 5, United States Code, of the rule submitted by the Office of the Comptroller of the Currency of the Department of the Treasury relating to the review of applications under the Bank Merger Act; providing for consideration of the joint resolution (S.J. Res. 31) providing for congressional disapproval under chapter 8 of title 5, United States Code, of the rule submitted by the Environmental Protection Agency relating to "Review of Final Rule Reclassification of Major Sources as Area Sources Under Section 112 of the Clean Air Act"; and waiving a requirement of clause 6(a) of rule XIII with respect to consideration of certain resolutions reported from the Committee on Rules.

Version:  Reported in House (rh)
Date:     2025-05-19
Congress: 119

Source: https://www.govinfo.gov/content/pkg/BILLS-119hres426rh/xml/BILLS-119hres426rh.xml

Sponsored-By: L000600